requirements.txt
python-dotenv
google-generativeai
pyupbit
pybithumb
ta
openpyxl
